requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
fastapi==0.115.0
filetype==1.2.0
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.12
langchain-google-genai==4.2.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-checkpoint-postgres==3.0.4
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.2
orjson==3.11.7
openai==1.55.3
ormsgpack==1.12.2
packaging==26.0
ply==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
seatable-api==3.1.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.0
uuid_utils==0.14.0
websockets==15.0.1
wsproto==1.3.2
xxhash==3.6.0
zstandard==0.25.0
